+++ app/src/main/java/org/breezyweather/common/ui/widgets/InkPageIndicator.kt
@@ -25,7 +25,6 @@ import android.content.Context
 import android.graphics.Canvas
 import android.graphics.Paint
 import android.graphics.Path
-import android.graphics.Rect
 import android.graphics.RectF
 import android.os.Parcel
 import android.os.Parcelable