+++ app/src/main/java/org/breezyweather/settings/compose/BackgroundUpdatesSettingsScreen.kt
@@ -40,7 +40,6 @@ import androidx.compose.ui.platform.LocalUriHandler
 import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.res.stringResource
 import androidx.core.net.toUri
-import org.breezyweather.BuildConfig
 import org.breezyweather.R
 import org.breezyweather.background.weather.WeatherUpdateJob
 import org.breezyweather.common.basic.models.options.UpdateInterval